move all propagator executable loaders to separate file
